+++ assets/default_input_bindings.json
@@ -11,7 +11,8 @@
                     "target": "player:flatview/interact/rotate",
                     "multiply_value": 0.01
                 }
-            ]
+            ],
+            "/mouse/click": "player:flatview/interact/press"
         }
     },
     "input:keyboard": {
@@ -88,6 +89,7 @@
     },
     "vr:controller_left": {
         "event_bindings": {
+            "/actions/main/in/press": "vr:left_hand/interact/grab",
             "/actions/main/in/grab": "vr:left_hand/interact/grab",
             "/actions/main/in/rotate_left": {
                 "target": "vr:origin/action/snap_rotate",
@@ -101,6 +103,7 @@
     },
     "vr:controller_right": {
         "event_bindings": {
+            "/actions/main/in/press": "vr:right_hand/interact/grab",
             "/actions/main/in/grab": "vr:right_hand/interact/grab",
             "/actions/main/in/rotate_left": {
                 "target": "vr:origin/action/snap_rotate",
