+++ assets/scenes/menu.json
@@ -50,12 +50,41 @@
 				"window": "lobby"
 			},
 			"screen": {
-				"target": "gui",
 				"luminance": [
 					3,
 					1,
 					3
 				]
+			},
+			"script": {
+				"prefab": "template",
+				"parameters": {
+					"source": "world_gui"
+				}
+			}
+		},
+		{
+			"transform": {
+				"translate": [
+					0,
+					1.4,
+					3.9
+				],
+				"rotate": [
+					180,
+					0,
+					1,
+					0
+				]
+			},
+			"gui": {
+				"window": "inspector"
+			},
+			"script": {
+				"prefab": "template",
+				"parameters": {
+					"source": "world_gui"
+				}
 			}
 		},
 		{
