+++ assets/scenes/templates/world_gui.json
@@ -1,0 +1,17 @@
+{
+    "components": {
+        "physics": {
+            "group": "interactive",
+            "dynamic": false,
+            "shapes": {
+                "box": [1, 1, 0.01]
+            }
+        },
+        "event_input": [
+            "/interact/point",
+            "/interact/press"
+        ],
+        "screen": {
+        }
+    }
+}
