+++ src/graphics/graphics/gui/WorldGuiManager.hh
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "GuiContext.hh"
+
+namespace sp {
+
+    class WorldGuiManager : public GuiContext {
+    public:
+        WorldGuiManager(ecs::Entity guiEntity, const std::string &name);
+
+        virtual void BeforeFrame();
+
+    private:
+        ecs::EntityRef guiEntity;
+    };
+} // namespace sp
